✨first pass support for recursive import checks
Looks at package.json using bin/main/files to figure out all the files that will be packed when the package is published then recursively crawls all `require`d files and dependencies.

Required files are not found in the list of packged files are surfaced as an error.

Does nothing with 3rd-party deps yet but can be used to find imports that are missing from the package.json.